+++ package.json
@@ -3,10 +3,17 @@
   "author": "Roland Zeng <roland.zeng@gmail.com>",
   "version": "0.0.1",
   "description": "Returns days until decay for League of Legends account",
+  "devDependencies": {
+    "react-scripts": "0.8.4"
+  },
   "dependencies": {
     "async": "^2.1.4",
+    "limiter": "^1.1.0",
     "moment": "^2.17.1",
     "moment-range": "^2.2.0",
+    "morgan": "^1.7.0",
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
     "request": "^2.79.0",
     "utf8": "^2.1.2"
   },
@@ -15,7 +22,11 @@
     "url": "https://github.com/rzeng95/LolDecay"
   },
   "scripts": {
-    "start": "node ./server.js",
-    "dev": "nodemon ./server.js"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "server": "node server",
+    "dev": "nodemon server"
   }
 }
