--- package_old.json
@@ -1,31 +1,0 @@
-{
-  "name": "LolDecay",
-  "author": "Roland Zeng <roland.zeng@gmail.com>",
-  "version": "0.0.1",
-  "description": "Returns days until decay for League of Legends account",
-  "dependencies": {
-    "async": "^1.5.2",
-    "body-parser": "latest",
-    "cron": "latest",
-    "express": "latest",
-    "iconv-lite": "latest",
-    "jade": "latest",
-    "limiter": "latest",
-    "moment": "latest",
-    "moment-range": "latest",
-    "mongoose": "latest",
-    "nodemailer": "latest",
-    "request": "latest"
-  },
-  "engines": {
-    "node": "0.12.7"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/rzeng95/LolDecay"
-  },
-  "scripts": {
-    "start": "node ./server.js",
-    "dev": "nodemon ./server.js"
-  }
-}
